% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_highmoments.R
\name{acf_highmoments}
\alias{acf_highmoments}
\title{Plot the higher moment auto-correlations and their cross-correlations}
\usage{
acf_highmoments(prices, on = c("days", "weeks", "months"), moments = c(1, 2,
  3, 4, 5))
}
\arguments{
\item{prices}{An xts of prices for the asset under analysis.  Only the
first column is used.}

\item{on}{A character vector specifying the return granularity to
calculate.  Valid granularities include any or all of:
"days", "weeks" and "months".}

\item{moments}{Specifies which moments of returns to calculate. The
autocorrelations and cross-correlations will be performed
for all permutations.}
}
\value{
No value is explicitly returned, but several plots are generated
        using the acf function from the stats package.
}
\description{
This function is useful to explore the correlation of returns within
a time series. A price series is given, and returns are calculated for
the specified periods.  Higher moments of returns are also calculated up
to the quintic returns (power of 5). Autocorrelations and cross-correlations
are then computed and plotted.
}
