% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_opt.R
\name{backtest_opt}
\alias{backtest_opt}
\title{Function to backtest a periodically optimized portfolio (empty)}
\usage{
backtest_opt(R, portfolio, optimize_method = "random", search_size = 100)
}
\arguments{
\item{R}{the returns}
}
\value{
return value
}
\description{
This function repeatedly calls function optimize.portfolio in PortfolioAnalytics
to perform a backtest of a strategy requiring optimization. Since optimize.portfolio
can be slow to run, it provides periodic updates at each optimization date via a log
file. It is assumed to run on multiple CPU cores using foreach and the snow package.
}
\details{
This function uses the version 2 of optimize.portfolio.  More details by
?optimize.portfolio.
}
\examples{
 backtest_opt()
}

