% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{cumsum_na}
\alias{cumsum_na}
\alias{cumprod_na}
\title{Compute cumulative sums or products and ignores NAs.}
\usage{
cumsum_na(x)

cumprod_na(x)
}
\arguments{
\item{x}{Value of the series from which to compute the cumulative sum or product.}
}
\value{
A numeric vector of the cumulative sum or product.
}
\description{
Functions cumsum_na() and cumprod_na() are similar to cumsum() and cumprod(),
except that they ignore any NAs present within the argument matrix.
}
\details{
NAs are ignored by assuming they equal zeroes for cumsum_na() and equal
ones for cumprod_na(). These functions work on vectors and xts matrices.
With an xts matrix, it applies cumprod or cumsum on every column.
}
\section{Functions}{
\itemize{
\item \code{cumprod_na}: Compute the cumulative sum while ignoring NAs
}}

\examples{
cumsum_na(c(NA, 1, 2, 1, NA, 0, 0))
cumprod_na(c(NA, 1, 2, 1.3, NA, 1, 2.5, NA))

# compute returns on xts_data, first row contains NAs
rets <- ROC(xts_data, type="discrete")
ecurves <- cumprod_na(1 + rets)
xtsplot(ecurves, main = 'Equity curves')
xtsplot(xts_data, main = 'Prices')   # identical since normalized.

}
