% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_momentum.R
\name{dual_momentum}
\alias{dual_momentum}
\title{Implements the dual momentum strategy as implied by Antonacci's paper.}
\usage{
dual_momentum(prices, assets, relmom, absmom)
}
\arguments{
\item{prices}{An xts of prices for all assets}

\item{assets}{A vector of the asset symbols to use as the module's
dual_momentum universe.  All symbols should correspond to
a column in the prices xts.  If an interest bearing cash
instrument is desired, then it should be included here also
and always produce a positive momentum via interest.}

\item{relmom}{Either a numeric or an xts containing the momentum
values for all assets used to compare between assets (relative).
If a numeric value, then it is assumed to be the momentum period
in days and a momentum matrix will be computed from the prices
matrix.}

\item{absmom}{Either a numeric or an xts containing the momentum
values for all assets used to apply the absolute momentum filter.
An asset must have positive momentum to be considered.
If a numeric value, then it is assumed to be the momentum period
in days and a momentum matrix will be computed from the prices
matrix.}
}
\description{
If no interest bearing cash instrument is provided, then it is possible
that all assets will have negative momentum.  In such case, the function will
go to cash and return no interest (daily returns = 0)
}
