% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{emptyxts}
\alias{emptyxts}
\title{Create an empty xts matrix filled with NAs}
\usage{
emptyxts(cnames = NULL, nc = 1, rowfill = NA,
  order.by = index(xts_data["2014", ]))
}
\arguments{
\item{cnames}{A character vector of column names, or NA
if not specified.}

\item{nc}{The number of columns unless cnames is provided,
which overrides this parameter.}

\item{rowfill}{A vector of numbers which is used to fill each
row of the xts matrix.  If the vector is too short,
it is recycled, or truncated if too long.  If NA, then
the xts will be empty (all NAs).}

\item{order.by}{The index of the xts matrix}
}
\value{
An empty xts matrix filled with NAs
}
\description{
Create an empty xts matrix filled with NAs
}
