% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_statfolio.R
\name{extract_SDweights}
\alias{extract_SDweights}
\title{Extract the SD weights of a statfolio for further processing}
\usage{
extract_SDweights(statfolio, tf = "all")
}
\arguments{
\item{statfolio}{The statfolio from which to extract the SDweights}

\item{tf}{The statfolio timeframe of interest, in xts index format}
}
\value{
Returns an xts containing the weights SD of each asset at each
        WFO date.
}
\description{
Extract the SD weights of a statfolio for further processing
}

