% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find outliers in a distribution and optionally removes them}
\usage{
find_outliers(dt, var)
}
\description{
Although removing outliers is normally a bad idea, in some
cases this can be useful when the interest lies in the shape
of the main distribution.  This function finds these outliers
and removes them if specified so.
}
\details{
This function needs to be properly documented...
}

