% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{flip_xtslist}
\alias{flip_xtslist}
\title{Flips the elements of a list of xts with the elements of the columns of each
underlying xts.}
\usage{
flip_xtslist(xtslist)
}
\description{
This function is used when a list of xts of similar structure is provided, but we
want to flip its organization.  For instance, if a list of xts contains 3 features
for 5 assets and is organized as follows:   A list of the 5 assets, and each element
contains an xts of the 3 features for that asset.
}
\details{
Invoking flip_xtslist will reorganize this list as follows:  A list of the 3 features,
and each feature will contain an xts of the 5 assets.

Note that the data is left untouched. Only its organization is affected.
}
