% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{fnamestamp}
\alias{fnamestamp}
\title{Extends a file name with a datetime stamp.}
\usage{
fnamestamp(fname = "NoName", asis = FALSE)
}
\arguments{
\item{fname}{Name of the file, with or without the extension.  The datetime
stamp will be inserted just before the extension, or at the end
if the file has no extension specified.  An extension is any
string after the last character dot in fname.}

\item{asis}{Logical.  If TRUE, the time stamp is appended at the end, even
if an extension exists.}
}
\value{
A character string made up of the filename, the datetime stamp and the
        extension (located either before or after the datetime stamp).
}
\description{
Returns a character string (the filename) with a datetime stamp appended for
easy output file naming.
}
\examples{
fnamestamp('./my cool file name.pdf')
fnamestamp('./my other cool file name.pdf', asis=TRUE)
}
