% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{getname}
\alias{getname}
\title{Get the name of an object}
\usage{
getname(object)
}
\arguments{
\item{object}{The object from which its name is sought.}
}
\value{
Returns the name of the object passed as argument, in the form of a string.
}
\description{
Get the name of an object
}
\examples{
myobject <- 5
getname(myobject)

}

