% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_returns.R
\name{interpolate_returns}
\alias{interpolate_returns}
\title{Interpolate returns in a time series containing NAs.}
\usage{
interpolate_returns(rets, value = c("prices", "returns"))
}
\arguments{
\item{rets}{The xts matrix of returns, which contains some NAs to be
interpolated.}

\item{value}{Specifies whether the return value is an xts of returns ("returns") or
a normalized equity curve "prices").  Default is "prices".}
}
\value{
Either an xts of returns or an xts of the equity curve, depending
        on the value argument.
}
\description{
An xts matrix of returns (1 or more columns is assumed).  The matrix
is first converted to an equity curve so intermediate prices can be
interpolated.
}

