% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdplot.R
\name{jdplot}
\alias{jdplot}
\title{Function to visualize the predictive power of a joint distribution}
\usage{
jdplot(x, y, target, mode = "simple", qtiles = list(Top = c(0.95, 1), Bottom
  = c(0, 0.05)), legendloc = "topleft")
}
\arguments{
\item{x}{The x and y data features.  These are normally 2 columns from an xts matrix.}

\item{target}{The target variable, also typically from an xts matrix column.}

\item{mode}{The plotting mode.  Supported modes are 'simple' for a simple scatterplot#                   to highlight the quantile regions using a given pch character and a given#                   color.}

\item{qtiles}{A named list containing the quantiles to analyze.  The quantile names can be}
}
\description{
Function to visualize the predictive power of a joint distribution
}
