% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Loads parameters from an excel file to spawn multiple parametric runs}
\usage{
load_parameters(fname, data = NA)
}
\arguments{
\item{fname}{The excel file name containing the parameters for each
parametric run.  Must include the proper file path.}

\item{data}{If data is provided, then it is used instead of loading
a file to prepare a dataframe of parametric runs. The
main utility of this is to do rules checking on the data
to ensure that all columns are present.  For example,
the data provided may have some missing columns but the
dataframe returned will have these columns filled with
appropriate default values.  Default is NA (read from file).}
}
\value{
Returns a dataframe, which each column properly cleaned up and
        containing the parameters for the runs.  Each row corresponds to
        a unique run.
}
\description{
Loads parameters from an excel file to spawn multiple parametric runs
}

