% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mltimer.R
\name{make_mltimer}
\alias{make_mltimer}
\title{Converts the output list of a machine learning predictor into a timer}
\usage{
make_mltimer(mlresults, prices, rules, trade_day = 1)
}
\arguments{
\item{mlresults}{The output list generated by a wfo machine learning
asset price predictor.}

\item{prices}{An xts matrix containing the price of the asset being
timed using the predictions of the machine learner. This
should be a one column xts matrix.}

\item{rules}{A list of equity curves to generate from the timer.  Each list
object consist of a named list which names the equity curve and
contains the rules to construct that equity curve.  The first
item on each named list must be the name of the ruleset to follow,
and other items are its arguments, which may be optional. For
details on the supported rules, see function
\code{\link{make_equitycurve}}.}

\item{trade_day}{Specified when the timer trade is executed. trade_day = 1 means
the trade is executed at the close of day when the timer is
calculated.  trade_day = 2 means at the next close, and so on.}
}
\value{
Returns a list containing the folloring xts matrices:
\describe{
  \item{\preformatted{$long}}{
     An xts matrix containing the equity curves associated with each timer method,
     computed daily, and long when the post-processed timer is positive, or
     in cash otherwise (without earning any interest). The normalized asset
     price equity curve is also included to enable an easy comparison.
  }
  \item{\preformatted{$timers}}{
     An xts matrix containing the binary timers associated with each timer method,
     computed daily, indicating one when the post-processed timer is positive, or
     zero otherwise.
  }
}
}
\description{
Converts the output list of a machine learning predictor into a timer
}
