% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_features.R
\name{make_transform}
\alias{make_transform}
\title{Make simple transformations of an xts matrix}
\usage{
make_transform(data, feature, addname = FALSE)
}
\arguments{
\item{data}{The xts matrix containing the data to transform. The
feature is applied to each column as specified by
argument `feature`.}

\item{feature}{The definition of the feature to apply to the data
matrix. See the Details section for `make_features` for
the syntax for specifying a simple feature.}

\item{addname}{Logical. Specifies whether the column name of the
xts matrix returned stay as they were provided, or
whether the feature name is appended to it.}
}
\value{
Returns an xts matrix of identical dimensions as data, with each
        column having the specified feature applied to it. Leading NAs are
        padded as needed when computing the feature.
}
\description{
This function takes an xts matrix and applies a simple
feature transformation to it in a columnwise fashion.  For example,
the matrix may be a series of asset prices.  The feature transformation
could be to perform a rolling SMA to it.
}
\seealso{
make_features
}
