% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxscreen.R
\name{maxscreen}
\alias{maxscreen}
\title{Volatility adjusted momentum screen for optimizing portfolios}
\usage{
maxscreen(prices, maxweights = 0.5, volfeature = "sd252",
  volthresh = 0.12, momfeature = "mom252", momthresh = 0.05,
  cashasset = 1)
}
\arguments{
\item{prices}{An xts matrix of asset prices.}

\item{maxweights}{A vector containing the maximum allowable weight
for each asset. The vector is recycled or truncated
if too short or too long.  Default is 0.5.}

\item{volfeature}{The volatility feature that will be computed from
the prices matrix using function make_feature.
Default is sd252.}

\item{volthresh}{The volatility threshold at which the asset weight
will be scaled back, according to the formula
min(1, volthresh / asset_volatility), where
asset_volatility is the result of the volfeature
feature.}

\item{cashasset}{The column number or name of the cash asset.  This
asset has a fixed weight of 1 so the optimizer can
always select it, if all other assets have a zero
weight.  Default is column 1.}
}
\value{
Returns an xts matrix of asset weights. Some weights may be
        zero if the asset was screened out at the given time index.
}
\description{
Volatility adjusted momentum screen for optimizing portfolios
}

