% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mget_symbolData.R
\name{mget_symbolData}
\alias{mget_symbolData}
\title{Get close price and additional data for multiple symbols.}
\usage{
mget_symbolData(symbols)
}
\arguments{
\item{symbols}{A vector of symbol names (character vector) to get symbol
data for.}
}
\value{
Returns a data frame where each row corresponds to a symbol, and
        each column corresponds to a specific data field found on the Yahoo
        page.
}
\description{
This function scrapes the yahoo quote page for certain types of information.
It gets the most recent close price and several other information related
to each symbol.  The information is compiled in a data frame where each
row is a symbol, and columns correspond to data related to the symbols.
}
\details{
Data fetched from the yahoo page include:
\itemize{
   \item{Symbol}
   \item{Inception Date}
   \item{Description on Yahoo}
   \item{Close Price}
   \item{Average Volume}
   \item{Net Assets}

}
}
