% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mget_symbols.R
\name{mget_symbols}
\alias{mget_symbols}
\title{Get multiple symbols and store in a single xts matrix}
\usage{
mget_symbols(symbols, from = "1999-01-01", OHLC = "Ad", src = "yahoo",
  filepath = "../DATABASE/data", ...)
}
\arguments{
\item{symbols}{A vector of symbol names to download (character vector).}

\item{from}{The start date from which the data is downloaded.}

\item{OHLC}{Specifies the value to be extracted from the OHLC data.
The default is "Ad" for adjusted close.  Must be
one of c("Ad", "Cl", "Op", "Hi", "Lo")}

\item{src}{Source where to find the data.  If src = "database", then
the symbol is searched in the local database with path
as specified in argument filepath.  Otherwise, this is
passed to quantmod::getSymbols function.}

\item{filepath}{The file path where to find the local database.  Default
is: "../DATABASE/data". This is used only if src = "database".
In that case, a file name will automatically be created
consisting of the symbol string concatenated with ".csv".}

\item{...}{Additional arguments passed to quantmod::getSymbols}
}
\value{
Returns an xts matrix starting at startdate or later if no symbol
        data is available at the earlier dates.
}
\description{
This function uses quantmod getSymbols to download multiple symbols
from a data source.  It then extracts one set of values (column) from
the OHLC data for each symbol and stores it into a wide xts matrix for
easy reference. Proper data columns are selected by invoking the appropriate
function from OHLC.Transformations in quantmod.
}
\details{
For example, if value = "Cl", then quantmod::Cl() function is invoked to
extract the Close column for each symbol.

If a symbol does not have enough history, leading NAs
are introduced for that symbol.  If the startdate provided is
earlier than all dates when symbol data is available, then theses dates
are omitted.
}
\examples{
x <- mget_symbols(c('SPY', 'QQQ', 'EWC', 'GLD'), startdate="1999-03-01")
head(x, 15)
y <- mget_symbols(c('SPY', 'QQQ', 'EWC', 'GLD'), startdate="1999-03-01", OHLC="Cl")
head(y, 15)
}

