% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirorbarplot.R
\name{mirrorbarplot}
\alias{mirrorbarplot}
\title{Function to create a mirrored bar plot.  This is two overlaid barplots,
one generally above the zero line and the other below the zero line.
A set of periodic returns is given (rets), along with a set of maximum
drawdowns (mdd).  The rets are plotted as positive bars (green) and the
mdd as negative bars (red). If the rets are negative, then they show up
as negative orange bars.}
\usage{
mirrorbarplot(rets, mdd, cex.lab = 1, barcolors = c("red", "green",
  "orange"), cex.axis = 1, cex.legend = 0.8, cex.names = 1,
  cex.main = 1, main = "Mirror Barplot", cex.text = 0.8,
  legend.loc = c(17, -25), yrange = NULL, xmgp = c(3, 1, 0),
  annotateYrs = TRUE, legend.horiz = FALSE, ...)
}
\arguments{
\item{rets}{An xts of periodic returns, typically annual returns.}

\item{mdd}{An xts of periodic maximum drawdowns with the same structure
as rets.}

\item{barcolors}{A vector of 3 colors for the bars.  Default red, green, orange.}

\item{main}{The title text.}

\item{legend.loc}{The X,Y location of the legend.  If NULL, then no legend is plotted.}

\item{yrange}{The numeric range for the Y axis.  Useful to extend / shorten it.
NULL will calculate it automatically.}

\item{xmgp}{See par() for mgp.  This moves the title and the axises around.}

\item{annotateYrs}{Logical. Whether to show the years on the X axis.  Note that the
period of returns should be years for this to work properly.}

\item{legend.horiz}{Logical. Passed on as horiz to legend() and used to set
the legend horizontally rather than vertically.}

\item{cex.<param>}{These are the various cex values.  cex.lab for the Y-label size,
cex.axis for the Y-axis size, cex.legend for the legend size,
cex.names for the X-axis label size, cex.main for the title,
cex.text for the annotated text on the plot (the average line).}
}
\value{
Nothing is returned per se.  This function creates a plot.
}
\description{
This is normally used to show annual returns for an equity curve overlaid
with its associated annual max drawdown.
}
