% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_features.R
\name{MSD}
\alias{MSD}
\title{Calculate the Mean-Squared Difference of a 2 column matrix}
\usage{
MSD(mat)
}
\arguments{
\item{mat}{The two column matrix used to calculate the mean squared difference.}
}
\value{
Returns the mean-squared difference of the two columns in the matrix.
}
\description{
Given a two column matrix (typically yhat and y), this function
calculates the mean-squared difference and returns the result.
}

