% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_statfolio.R
\name{optimize_statfolio}
\alias{optimize_statfolio}
\title{Optimize a portfolio of assets N times to get asset weight statistics}
\usage{
optimize_statfolio(rets, portfolio, train_window = 63, N = 1,
  weightFUN = "mean", objfnmat = NA, maxwtsmat = NA,
  optimize_method = "DEoptim", ...)
}
\arguments{
\item{rets}{An xts matrix of asset returns.}

\item{portfolio}{An object of type "portfolio" specifying the constraints
and objective function.}

\item{train_window}{The training window (in days) used to subset rets to calculate
the optimization matrix. This is normally the same as the
rolling_window if WFO optimization is used.}

\item{N}{The number of times to repeatedly call function
optimize.portfolio to generate weight statistics.
Default is 1.}

\item{weightFUN}{Sets the method used to compute the asset weights returned
from all runs of optimize.portfolio.  Default is "mean",
but it can be any valid function name such as "median" or other.}

\item{objfnmat}{The feature matrix passed to the custom objective function, if
specified, used by optimize.portfolio.}

\item{maxwtsmat}{The maximum weights xts matrix.  It should include a row with
the current optimization date, which will be extracted as a vector
and added as a constraint in the optimize.portfolio function call.
Ignored if not specified.}

\item{optimize_method}{Sets the optimization method used by optimize.portfolio.
Default is "DEoptim".}

\item{...}{Additional arguments passed through to optimize.portfolio.}
}
\value{
Returns a list containing the same elements as what optimize.portfolio
         normally returns, with the following exceptions:

\describe{
  \item{\preformatted{$weights}}{
     A named vector containing the optimal set of weights for the portfolio.
     If N > 1, then this is a statistic computed from all optimize.portfolio
     runs.  The choice of statistic is a function named using argument weightFUN.
     Default is "mean".
  }
  \item{\preformatted{$N}}{
     The number of times function optimize.portfolio was called.
  }
  \item{\preformatted{$allweights}}{
     A matrix containing the optimized asset weights for each run of
     optimize.portfolio
  }
  \item{\preformatted{$SD_weights}}{
     A named vector containing the standard deviations of the weight values,
     computed by taking the StdDev of $allweights.  If N = 1, then this
     will be all zeroes.
  }
}
}
\description{
Optimze a portfolio of assets N times in order to gather statistics
on the portfolio asset weights.  The optimization is performed at the
most recent date as provided by the xts matrix of returns.
}
\details{
This function leverages the optimize.portfolio function from package
PortfolioAnalytics. It differs from optimize.portfolio in
several ways. First, an argument N is provided to repeatedly call
optimize.portfolio in order to gather statistics on the asset weights.
This is relevant because often, the optimization of the objective function
results in a somewhat unstable optimum, resulting in many possible asset
weights.

See the vignette xtsanalytics for more details on the statfolio data structure.
}
\seealso{
optimize.portfolio
}

