% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_mlpar.R
\name{pad_mlpar}
\alias{pad_mlpar}
\title{Pads a list of ML algorithm paramters with defaults if not provided}
\usage{
pad_mlpar(mlalgo = NULL, mlpar = NULL)
}
\arguments{
\item{mlalgo}{The name, in quote or the ml algorithm to populate with
its parameters.  Supported ml algorithms include "h2o.rf",
"xgboost".}

\item{mlpar}{A list of named ml parameters to override the default. If
nothing is specified, then NULL is assumed and the funtion
use all the embedded defaults.}
}
\value{
Returns a list of parameters for the said ml algorithm.
}
\description{
For any given ML algorithm, several parameters must be provided.  This
function will return a list (mlpar) with a sensible set of default
arguments, unless it is overridden by specifying it in the mlpar
argument.
}

