% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perfstats}
\alias{perfstats}
\title{Compute performance statistics on equity curves}
\usage{
perfstats(data, on = "days", plotout = TRUE, value = TRUE,
  percent = TRUE, digits = 2, top = 3, main = "Performance Summary",
  scaling = "auto", title_size = "auto", ...)
}
\arguments{
\item{data}{An xts matrix of equity curves, each column being one equity curve.}

\item{on}{The periodicity (scale) of the xts data matrix to compute the stats.
"days", "weeks", "months", "quarters" and "years" are valid
and correspond to the following scales for table.AnnualizedReturns
respectively:  252, 52.18, 12, 4 or 1.}

\item{plotout}{Logical.  Indicates whether to plot the result on the graphic console.
using function plot_df.  Default is TRUE.}

\item{value}{Logical (default TRUE) to indicate whether a dataframe is returned}

\item{percent}{Logical to indicate whether to multiple by 100 to get percentages
(default TRUE).}

\item{digits}{Number of digits to show on the screen / file and returned dataframe.}

\item{top}{The number of drawdowns to calculate, starting from the maximum drawdown.}

\item{main}{The title of the table to plot on the graphic console.  Default is
"Performance Summary".}

\item{scaling}{Table scaling factor to plot on the graphic console.  Default is "auto",
which automatically calculates a reasonable value based on the
number of columns in the xts data matrix.}

\item{title_size}{Title scaling factor passed on to plot_df.  Default is "auto",
which calculates a reasonable value based on the number of
coumns in the xts data matrix.}

\item{...}{Additional arguments passed on to plot_df.  For example, the title
text and size, and footnote size may be modified using this.}
}
\value{
A dataframe that shows the summary performance of the equity curves.  The
       annualized return, the standard deviation, the Sharpe ratio and the maximum
       drawdown.
}
\description{
Returns a dataframe containing performance stats on one or multiple
equity curves.  May also plot it on the graphic device.
}

