% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification stats.R
\name{periods_in_class}
\alias{periods_in_class}
\title{Cumulatively counts the periods since the latest prediction}
\usage{
periods_in_class(data, col = 1, retcol = "cum_periods")
}
\arguments{
\item{data}{The xts matrix data containing the classifier's results column
(e.g. market timer).}

\item{col}{The number or name of the column containing the classification information.}

\item{retcol}{The name of the column containing the cumulative count results
(default is "cum_count").}
}
\value{
A column added to the xts matrix data argument.
}
\description{
This function computes the cumulative number of periods since the last prediction
(or transition).  For instance, a market timer is a form of binary classifier and
when applied to the timer series, this function will return the number of periods
since it last changed value.
}
\details{
The return value is in the form of a column added to the xts matrix passed as the
data argument.  The added column contains the cumulative period count since the last
observed transition in the specified predictor column.
}

