% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameters.R
\name{plot_parameters}
\alias{plot_parameters}
\title{Plot the parameters for a simulation run}
\usage{
plot_parameters(data, symbols = NA, nc = 8, main = "Run Parameters",
  symbol_title = "List of Asset Symbols", exclude_rows = NA)
}
\arguments{
\item{data}{Either a list of parameters, or a dataframe containing
the run parameters to plot on the screen.  If a list,
each item should be named as the names correspond to the
parameter name.  If a dataframe, the row names should
be the parameter names.}

\item{symbols}{A character vector of the asset ticker symbols.  This
set of assets will be printed near the bottom of the
screen in a matrix format of up to 8 columns and as
many rows as needed. If omitted (NA), then nothing is
printed.}

\item{nc}{The maximum number of columns to display the set of assets.
Default is 8.}

\item{main}{The main title at the top panel of the screen.  If no
title is desired, then set to an empty string "".}

\item{symbol_title}{The title for the bottom panel where the asset
symbols are shown.}

\item{exclude_rows}{A vector of row numbers or row names corresponding
to the rows to exclude from showing in the top panel.}
}
\description{
Plots the parameters for a simulation in its own plotting window.
This is normally used as a header page to describe the key
parameters for a simulation.  The function provides
}

