% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rsquare.R
\name{plot_rsquare}
\alias{plot_rsquare}
\title{Scatterplot of a regressed predictor vs. actual to visualize ML model performance}
\usage{
plot_rsquare(xy, main = NA)
}
\arguments{
\item{xy}{A 2 column xts matrix containing y and ypred.}

\item{main}{If provided, overrides the default plot title}
}
\value{
The linear regression model is returned
}
\description{
This function does a scatter plot of a predicted value (ypred) vs. its
 actual value (y). The points are color coded such that if the signs of y and
 ypred are the same, the points are green, otherwise they are red.
}
\details{
In addition, a linear regression is performed and the regression line is shown
 in red, along with the R-square value in the subtitle.  A line is also plotted in
 dotted blue to illustrate the split of data at x = y.

NOTE:  This function assumes that column 1 is y, and column 2 is ypred.
}

