% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_statfolio.R
\name{plot_statfolio}
\alias{plot_statfolio}
\title{Plot equity curves and sensitivity information related to a statfolio}
\usage{
plot_statfolio(statfolio, tf = "all", type = "SDweights", Ncurves = 100,
  main = "Statfolio Performance Summary", log = "y", ...)
}
\arguments{
\item{statfolio}{A statfolio of class "optimize.portfolio.rebalancing" as
created by a WFO run using function wfo_statfolio.}

\item{tf}{The statfolio timeframe to consider, using standard xts
rules and formatting.  If tf = "all" (default) then the
entire available timeframe is used.}

\item{type}{The type of plot to generate.  Available types include
"SDweights" (default), which generates three equity curves
(best, average and worst) based on the weight variations
resulting from the WFO runs on the top panel.  On the bottom
panel, a time plots of the standard deviation of each asset
weight is shown.  This helps visualize whether the optimizer
converges to a stable set of weights (low SD) and whether that
variation, not matter how small, results in meaningful changes in
portfolio performance over time.}
}
\value{
Nothing is returned, except that a plot is generated on the graphic device.
}
\description{
Plot equity curves and sensitivity information related to a statfolio
}

