% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngfile.R
\name{pngfile}
\alias{pngfile}
\title{Function to ease the creation of a png graphic file}
\usage{
pngfile(onoff, outflag, fname = NA, fpath = NA, width = 1600,
  height = 1000)
}
\arguments{
\item{onoff}{A character string to denote whether the file in opened
"on" or closed "off". This must be specified.}

\item{outflag}{Specifies whether the graphical output is sent to the
console or the file.  This is useful in a script to
easily control where the output is sent when several
charts are created (send them all to file or the console).}

\item{fname}{The file name for the output file.}

\item{fpath}{The path of the output file.}

\item{width}{The width (in pixels) of the chart.}

\item{height}{The height (in pixel) of the chart.}
}
\value{
Nothing is returned per se.  This function opens/close the
        output channel.
}
\description{
This function is used to open the png file or to close the png file.
One opened, plots normally sent to the console will be sent to the
png file instead.
}
