% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_cor.R
\name{portfolio_cor}
\alias{portfolio_cor}
\title{Calculate pairwise asset correlation vs. a portfolio}
\usage{
portfolio_cor(rets, portassets = colnames(rets)[1], maxwtsvec = 1,
  on = "days")
}
\arguments{
\item{rets}{An xts matrix of asset returns that include all assets
in the analysis universe, the latter being a superset
of the portfolio assets.  The correlation is calculated
over all returns provided by rets.}

\item{portassets}{A character vector containing the names of assets in
the portfolio. Must contain at least one name.}

\item{maxwtsvec}{Either an ordered or named numeric vector containing the
relative weight of each asset in the universe or a single
number. The sum of all weights does
not need to equal one (it would normally exceed one), as
this is used for relative weights.  If a single number
(for example set to 1) then, this result in equal weights.}

\item{on}{Specifies the returns period to use to calculate the
correlation.  Default is "days", but can also be "weeks"
or "months" for weekly and montly correlations.  The portfolio
is also rebalanced at that same frequency.}
}
\value{
Returns a named vector of correlation values for each asset vs.
          the given portfolio.
}
\description{
Calculates the pairwise correlation between all assets in a universe
vs. a portfolio of assets.
}
\details{
The portfolio must comprise assets in the universe otherwise an error
will result.  A named vector is return with the correlation of each
asset not in the portfolio vs. the portfolio.  The portfolio may be
built of equal weights or a set of predefined weights, as specified
by maxwtsvec.
}

