% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification stats.R
\name{predictor_cor}
\alias{predictor_cor}
\title{Analyzes correlations between an asset and a predictor (EMPTY)}
\usage{
predictor_cor(data, price_col = "prices", timer_col = "GC",
  plot = TRUE)
}
\arguments{
\item{data}{The xts matrix data containing the equity curve of the asset or
portfolio, and the classifier's prediction column. e.g. market timer.
The predictor column may be binary (-1, +1) or real-valued centered
about zero.}

\item{price_col}{The column name or column number of the equity curve under test.
Default = "prices".}

\item{timer_col}{The column name or column number of the prediction (or timer).
Default is "GC".}

\item{plot}{Logical value indicating whether or not to output a scatter plot
annotated with the class correlations and a legend.}
}
\value{
Returns the class correlations and other information as a list.
\describe{
  \item{\preformatted{$pred_N_up:}}{
     Number of up predictions by the predictor.
     }
  \item{\preformatted{$actual_N_up:}}{
     Number of actual up observations of the asset.
     }
  \item{\preformatted{$up_cor:}}{
     The predictor correlation for the up class (predictor value is positive).
     }
  \item{\preformatted{$pred_N_down:}}{
     Number of down predictions by the predictor.
     }
  \item{\preformatted{$actual_N_down:}}{
     Number of actual down observations of the asset.
     }
  \item{\preformatted{$down_cor:}}{
     The predictor correlation for the up class (predictor value is negative).
     }
}
}
\description{
Computes the class correlations between an asset (or any equity curve) and
a predictor.  The predictor can be a binary class or a real value centered
about 0.
}
\details{
In the real valued case, it is assumed that the sign represents the
direction of the market, while the real value is a conviction level.  For example,
a predictor value of 0.0001 could be considered a weak conviction of an up market,
whereas a predictor value of -0.1 could be construed as a strong conviction of a
down market.  Person linear correlations (function cor) is used to compute
the correlation between the actual future market direction (internally represented
as +1, -1), and the real value of the indicator
}
