% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{rawreturns}
\alias{rawreturns}
\title{Compute the raw returns on an xts of prices}
\usage{
rawreturns(prices)
}
\arguments{
\item{prices}{The xts matrix of prices}
}
\value{
Returns the raw returns (not annualized) returns
of each xts series in percentage.
}
\description{
Compute the raw returns on an xts of prices
}
