% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcolors.R
\name{Rcolors}
\alias{Rcolors}
\title{Plots a map of the default colors for easy reference}
\usage{
Rcolors()
}
\description{
This will produce two plots:  the first one plots the colors
by name, and the second one by their respective Hex codes.
}
\details{
The code is courtesy of the blog New Energy Research at
bxhorn.com
}

