% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{recycle}
\alias{recycle}
\title{Recycle a data object and truncate to a length of N.}
\usage{
recycle(data, N)
}
\arguments{
\item{data}{Object containing the data to recycle}

\item{N}{The target length of the object once recycled.}
}
\value{
Returns the same object as provided, recycled enough times
        and truncated to ensure it has length N.
}
\description{
Recycle a data object and truncate to a length of N.
}
