% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{recycle_better}
\alias{recycle_better}
\title{Recycle a vector using names and smart rules}
\usage{
recycle_better(vec, vecnames, default = 0)
}
\arguments{
\item{vec}{A numeric vector used for recycling.  If the vector is
not named, then it is simply recycled N times where
N = length(vecnames). If it is a named numeric, then
a vector of length N is built using the default argument
below, then those named values are inserted in the
result.}

\item{vecnames}{The ordered names to assign to the results.}

\item{default}{The default value to pad the vector with when a named
vec argument is provided.}
}
\value{
Returns a vector of length(vecnames) that is named and ordered
        as vecnames.
}
\description{
Recycle a vector using names and smart rules
}
