% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollcor.R
\name{rollcor}
\alias{rollcor}
\title{Rolling correlation of one or multiple features vs. a target variable}
\usage{
rollcor(features, target, width = 63)
}
\arguments{
\item{features}{The xts matrix containing the features.}

\item{target}{An xts matrix with identical indices as the features matrix.
If the target is a prediction further out in time, then
the time relationship between features and target should
be properly lagged prior to calling this function.}

\item{width}{The rolling window width used to calculate the rolling
correlation between the features and the target.}
}
\value{
Returns an xts matrix of the same width as the feature matrix
         with the rolling correlation values.
}
\description{
Over a sliding window, calculate the rolling correlation of an xts
matrix of features vs. a target variable.
}
