% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{save_Rdata}
\alias{save_Rdata}
\title{Save an object to an .Rdata file with dateTime stamp and file path.}
\usage{
save_Rdata(object = NULL, fname = "temp", path = "./")
}
\arguments{
\item{object}{An object in quoted strings, or a list of objects,
found in the parent scope.  Can be an environment
or a variable.}

\item{fname}{File name}

\item{path}{File path, which usually starts with ./ and ends with /}
}
\value{
Side effect of saving the object to the specified file.
}
\description{
Save an object to an .Rdata file with dateTime stamp and file path.
}

