% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{split_vector}
\alias{split_vector}
\title{Split a vector into chunks of size k.}
\usage{
split_vector(vec, k)
}
\arguments{
\item{vec}{The vector to split into chunks.}

\item{k}{The size of each chunks.}
}
\value{
A list of vectors, each vector of size k, except possibly the last one.
}
\description{
This functions splits a vector into chunks of size k, and returns a list of
vectors.  The last vector in the list will be less than length k if k doesn't
divide the length of the vector exactly.
}
\details{
Parameter k must be of type numeric.  If not an integer, it will be coerced
to an integer using trunc().

Parameter vec must be a vector but can contain any element type.
}
\examples{
split_vector(c(1,2,3,4,5), 3)
}
