% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{splot}
\alias{splot}
\title{Simple scatterplot of two xts matrices with linear regression.}
\usage{
splot(x, y, tf = "2013/2019")
}
\arguments{
\item{x}{Scatterplot X axis expressed as a single column of an xts matrix}

\item{y}{Scatterplot Y axis expressed as a xts matrix column of the same
length as x.}

\item{tf}{The timeframe to use for the scatterplot}
}
\value{
This function creates a plot and does not return anything.
}
\description{
Performs a scatterplot of two xts matrix columns and does a linear
 regression on the results.
}
