% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsparallel.R
\name{start_parallel}
\alias{start_parallel}
\title{Starts the parallel backend.}
\usage{
start_parallel(Ncores = -1)
}
\arguments{
\item{Ncores}{A positive number specifies the number of cores to use by
the parallel backend.  0 means all cores.  A negative number
means all cores minus that number, so that the computer doesn't
"freeze up" while processing. Default is -1, so that one CPU core
remains unused.}
}
\value{
The parallel CPU cluster is returned and should be assigned to a variable
}
\description{
Starts the parallel backend cluster for use by the foreach function.
}
\seealso{
stop_parallel
}

