% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{startpoints}
\alias{startpoints}
\title{Finds the beginning of a period in an xts matrix}
\usage{
startpoints(x, on = "months", offset = 1, k = 1)
}
\arguments{
\item{x}{an xts matrix}

\item{on}{the string name for the period to extract (like endpoints)}

\item{offset}{the offset from the start of the period (offset = 1 at start date)}

\item{k}{how many periods to skip (k=1 means none are skipped)}
}
\value{
Returns a vector of the row positions in the matrix
}
\description{
This function is similar to endpoints except that it finds the start of the
period rather than the end of period.
}
