% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_dfdates.R
\name{subset_dfdates}
\alias{subset_dfdates}
\title{Subsets a dataframe by dates}
\usage{
subset_dfdates(df, timeframe, datecol, keep_NAs = FALSE)
}
\arguments{
\item{df}{The dataframe to subset by dates}

\item{timeframe}{A character string specifying the timeframe as two
full dates separated by a slash.  For example,
"2010-10-01/2010-10-22".}

\item{datecol}{The column name where the dates are found in the dataframe.
Although these may be of class char, they will be converted
to class Date before subsetting is applied.}

\item{keep_NAs}{Logical flag specifying whether to apply complete.cases
to the dataframe to eliminate rows with NAs.}
}
\value{
Returns a subsetted dataframe containing rows with dates within the
         timeframe supplied.  Also, complete.cases is applied to eliminate
         rows with NAs if keep_NAs = FALSE (default).
}
\description{
The dataframe argument must contain a column with valid dates.  The
timeframe argument specifies a timeframe xts-style daily.  For example,
"2011-01-01/2011-12-31" would be value whereas "2011" is not.
}

