% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_statfolio.R
\name{subset_statfolio}
\alias{subset_statfolio}
\title{Subset a statfolio using standard xts index formatting}
\usage{
subset_statfolio(statfolio, tf = "all")
}
\arguments{
\item{statfolio}{The statfolio to subset}

\item{tf}{The timeframe specified in standard xts format.}
}
\value{
A new statfolio that is a time subset of the original statfolio.  The
         daily returns xts matrix $R is not subsetted.  Only the $op_rebalancing
         portion is subsetted.
}
\description{
Subset a statfolio using standard xts index formatting
}

