% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_overnight.R
\name{trade_overnight}
\alias{trade_overnight}
\title{Builds an equity curve based on a successive number of overnight trades}
\usage{
trade_overnight(predmat, maxposn = 1, maxweight = 0.2, longthresh = 0.01,
  shortthresh = -100, dateseries = NA, datecol = NA, yhatcol = "yhat",
  retcol = "rets")
}
\arguments{
\item{predmat}{A dataframe or xts matrix containing a date column, a
prediction column (yhat) and an overnight returns colum.
The names of these columns should be specified with the
datecol, yhatcol and retcol arguments.}

\item{maxposn}{The maximum number of equally weighted simultaneous positions
to take for an overnight trade.}

\item{maxweight}{The maximum weight given to any trade.  This is important
to limit how large a position will be when only one or very
stocks are traded during one day.}

\item{longthresh}{The threshold above which a long trade is taken.  This
threshold is compared with the yhatcol value for this decision.}

\item{shortthresh}{The threshold below which a short trade is taken.  This
threshold is compared with the yhatcol value for this decision.
NOT YET IMPLEMENTED!}

\item{dateseries}{An optional xts series (such as SPY or GSPC) that has all dates
the market is open.  This is used to fill in dates not present
in predmat to carry forward the equity curve.  This way, the
resulting equity curve can be plotted without any big
time frames missing. Default is NA (no dates provided)}

\item{yhatcol}{The column name in predmat with the predicted values (yhat).}

\item{retcol}{The column name in predmat with the overnight returns.  Note
that the assumed overnight return starts on the current date
(close of market) to the next day's sale.  The user must
ensure these returns are properly lagged to ensure there is
no lookahead bias.}

\item{datacol}{The column name in predmat containing the dates.  If
predmat is an xts matrix, this should be set to NA which
will result in taking the xts matrix dates instead.}
}
\value{
Returns an xts matrix with the resulting equity curve.  The xts matrix
         dates will either correspond to the predmat dates only (dateseries = NA),
         or an outer merge of the predmat dates with the dates in dateseries,
         properly subsetted at the start and ending boundaries of predmat.

         In addition to the equity curve, the matrix includes a column for the
         number of trades performed on each date (Ntrades), and the total number
         of stocks available to trade on each date (Allstocks).
}
\description{
A dataframe with a date column is provided.  Each row corresponds to a
potential stock trade, so there may be many rows with different stocks on
the same date.  At least two columns must be provided:  the overnight
return, and an estimate (yhat) of that return generated by some model.
}

