% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{trimspaces}
\alias{trimspaces}
\title{Remove leading or trailing white spaces from a character string}
\usage{
trimspaces(x, type = c("both", "leading", "trailing"))
}
\arguments{
\item{x}{The character string to process}

\item{type}{Specifies whether to remove the leading, trailing or
both white spaces.}
}
\value{
Return the string provided without leading or trailing white
        spaces.
}
\description{
Remove leading or trailing white spaces from a character string
}

