% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ulcerindex}
\alias{ulcerindex}
\title{Calculate the ulcer index of equity curves}
\usage{
ulcerindex(data, type = "ec", on = "days", DDthresh = 0)
}
\arguments{
\item{data}{An xts matrix of at least one equity curve.}

\item{type}{Specifies whether returns ("rets") or an equity
curve ("ec") is provided for the data.}

\item{on}{Specifies whether the data matrix is sampled
on "days", "weeks", "months", "quarters" or "years".}

\item{DDthresh}{The drawdown minimum threshold for each
period's return to be considered in the
ulcer index calculation.  In other words,
drawdowns smaller than this number do not count.
Default is zero and may be expressed as
either a positive or negative number.}
}
\value{
Returns the Ulcer Performance Index for each
        equity curve provided.
}
\description{
Calculates the ulcer index of one or multiple
equity curves provided as an xts matrix.
}
\details{
The ulcer index is implemented according to the formula
described in the Wikipedia article "Ulcer Index".
}
