% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ulcerperformance}
\alias{ulcerperformance}
\title{Calculate the Ulcer Performance Index of equity curves}
\usage{
ulcerperformance(data, type = "ec")
}
\arguments{
\item{data}{An xts matrix of at least one equity curve.}

\item{type}{Specifies whether returns ("rets") or an equity
curve ("ec") is provided for the data.}
}
\value{
Returns the Ulcer Performance Index for each
        equity curve provided.
}
\description{
Calculates the Ulcer Performance Index of one or multiple
equity curves provided as an xts matrix.
}
\details{
UPi is implemented using the formula described in the
Wikipedia article "Ulcer Index" and UPI.
}

