% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xtsanalytics.R
\name{update_xtsanalytics}
\alias{update_xtsanalytics}
\title{Function to update xtsanalytics from github}
\usage{
update_xtsanalytics()
}
\description{
A simple function to update package xtsanalytics from
github if a new version is available.
}
