% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfo_grow_jungle.R
\name{wfo_grow_jungle}
\alias{wfo_grow_jungle}
\title{Parallel version of wfo_grow_forest}
\usage{
wfo_grow_jungle(featuremat, modelwindow = 252, wfo_span = "months",
  ylag = 21, Nblocks = 7, SPwindow = 63, verbose = FALSE, mtry = 2,
  ntree = 1000, importance = TRUE, na.action = na.omit, ...)
}
\description{
TODO write the documentation!!!
}
