% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfo_statfolio.R
\name{wfo_statfolio}
\alias{wfo_statfolio}
\title{Optimizes a portfolio of assets using walk-forward optimization}
\usage{
wfo_statfolio(rets, portfolio, roll_window = 63, N = 1, Ncores = 1,
  wfo_span = "months", weightFUN = "mean", objfnmat = NA,
  maxwtsmat = NA, optimize_method = "DEoptim", ...)
}
\arguments{
\item{rets}{An xts matrix of asset returns.}

\item{portfolio}{An object of type "portfolio" specifying the constraints
and objective function.}

\item{roll_window}{The rolling window (in days) used to subset rets to calculate
the optimization matrix.}

\item{N}{The number of times to repeatedly call function
optimize.portfolio to generate weight statistics at
each wfo_span points in time.  Default is 1.}

\item{Ncores}{The number of CPU cores used to run the WFO optimization in
parallel.  When set to 1 (the default), the code is single-threaded.
This does not automatically means only one core is used as some
underlying R functions may be multi-threaded by design.
When Ncores > 1, then the code will spawn a parallel backend and
use the foreach package to run its parallel jobs.  Foreach prevents
any outputs to the console, so the only way to see progress via
the console is to set Ncores = 1.}

\item{wfo_span}{The frequency (dates) when the portfolio is reoptimized and
rebalanced.  This can be set to any of: "weeks", "months", "quarters"
or "years".  Can be overridden with specific dates if such dates
are provided as a character vector of valid dates.}

\item{weightFUN}{Sets the method used to compute the asset weights returned
from all runs of optimize.portfolio.  Default is "mean",
but it can be any valid function name such as "median" or other.}

\item{objfnmat}{The feature matrix passed to the custom objective function, if
specified, used by optimize.portfolio.}

\item{maxwtsmat}{The maximum weights xts matrix, indexed by the WFO dates
so that maximum weights at these dates can be extracted before
calling optimize_statfolio.  The max weights are specified as
an added constraint in the optimize.portfolio function call.
Ignored if not specified.}

\item{optimize_method}{Sets the optimization method used by optimize.portfolio.
Default is "DEoptim".}

\item{...}{Additional arguments passed through to optimize.portfolio.}
}
\description{
This function repeated calls PortfolioAnalytics::optimize.portfolio at
a set of pre-defined dates to achieve walk-forward optimization. It is
similar to PortfolioAnalytics::optimize.portfolio.rebalancing except that
it allows for custom objective functions and it is possible to modify the
asset universe at each WFO date dynamically. This is useful when, for example,
the asset universe is screened for certain criteria prior to optimizing the
portfolio.
}

