% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfo_summary.R
\name{wfo_summary}
\alias{wfo_summary}
\title{Create a summary of key statistics from a Walk-Forward Optimization (WFO) run}
\usage{
wfo_summary(wfodata)
}
\arguments{
\item{wfodata}{The data structure containing the walk-forward optimization
information.  It should be of the proper class to run the
right underlying code generatic the summary.}
}
\value{
Returns an appropriate data structure summarizing the WFO run under
        analysis.  This normally includes an equity curve and, if a statfolio
        was provided, it would also include additional times series that quantifies
        the statistical properties of the set of equity curves resulting from the
        statfolio (details TBD).
}
\description{
Create a summary of key statistics from a Walk-Forward Optimization (WFO) run
}
\seealso{
wfo_statfolio
}

