% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfo_timeframe.R
\name{wfo_timeframe}
\alias{wfo_timeframe}
\title{Calculate the WFO dates and subset the feature matrix}
\usage{
wfo_timeframe(featuremat, modelwindow = 252, wfo_span = "months",
  ylag = 21, earliest = NA, latest = NA, verbose = FALSE)
}
\arguments{
\item{featuremat}{An xts matrix containing the features for the WFO model}

\item{modelwindow}{The size, in days, for the WFO model training window.}

\item{wfo_span}{The span for each WFO model or a character vector of dates.
Valid spans include "days",
"weeks", "months" or "quarters".  Uses function endpoints()
to calculate the model dates.}

\item{ylag}{An additional offset to add to the model window to adjust
for the amount of time the predicted variable (y) is lagged
in the WFO model.}

\item{earliest}{The earliest possible wfo date desired.  This is useful if
the featuremat goes much further back in time than needed.}

\item{latest}{The latest date wfo or a prediction is desired.  This is used
to truncate the wfo dates and also as the timeframe last date.}

\item{verbose}{If TRUE, the wfo_span dates will be printed on the console.}
}
\value{
Returns a list of the following elements:
\describe{
  \item{\preformatted{$wfo_dates}}{
     A character vector containing all WFO dates.
     }

  \item{\preformatted{$wfo_points}}{
     A numeric vector containing all WFO indices, as referred to featuremat.
     }

  \item{\preformatted{$wfo_start_date}}{
     The first WFO date.  Identical to $wfo_dates[1].
     }

  \item{\preformatted{$wfo_last_date}}{
     The last WFO date, which should also be the last available date in featuremat.
     }

  \item{\preformatted{$wfo_timeframe}}{
     The WFO timeframe, which is $wfo_start_date up to and including $wfo_last_date.
     Do NOT use this to subset the featuremat!
     }

   \item{\preformatted{$subset_timeframe}}{
     The timeframe that can be used to subset the featuremat matrix.  The first date
     is properly adjusted using modelwindow and ylag.  The last date is identical to
     $wfo_last_date.
     }

   \item{\preformatted{$Nlast}}{
     This the numeric index associated with the wfo_last_date. It can be used
     when numeric indexing on featuremat is preferable.
     }

   \item{\preformatted{$warn_msg}}{
     A character string containing a warning message if the earliest date cannot
     be met due to lack of data in featuremat.
     }

  }
}
\description{
Calculate the WFO dates and subset the feature matrix
}

