\name{xtsanalytics-package}
\alias{xtsanalytics-package}
\alias{xtsanalytics}
\docType{package}
\title{
\packageTitle{xtsanalytics}
}
\description{
\packageDescription{xtsanalytics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{xtsanalytics}
\packageIndices{xtsanalytics}

   The xtsanalytics package provides a set of functions to analyze xts
   time series more effectively. It leverages other packages including
   quantmod, PortfolioAnalytics, PerformanceAnalytics
   and others. The functions provides are
   meant to be used in the analysis of investment strategies and portfolio
   optimization.
}
\author{
\packageAuthor{xtsanalytics}

Maintainer: \packageMaintainer{xtsanalytics}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
