% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsanalytics.R
\docType{package}
\name{xtsanalytics}
\alias{xtsanalytics}
\alias{xtsanalytics-package}
\title{xtsanalytics: Portfolio analysis using optimization and machine learning.}
\description{
The xtsanalytics package provides a set of functions to analyze xts
time series more effectively for investment purposes. It leverages
several other packages including quantmod, PortfolioAnalytics,
PerformanceAnalytics and some machine learning packages including
h2o and xgboost.  The package is useful for developing portfolio
optimization strategies and predictive models using machine learning.
}
\section{Functions in xtsanalytics}{


Click on the Index below to see the list of available functions.
}

