% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsoverlay.R
\name{xtsbind}
\alias{xtsbind}
\title{Binds xts columns from two or more xts matrices.}
\usage{
xtsbind(x, y, ...)
}
\arguments{
\item{x}{First xts matrix to bind.}

\item{y}{Second xts matrix to bind.}

\item{...}{Additional xts matrices to bind.}
}
\value{
Returns a an xts matrix that cbinds all matrices provided as
        arguments, padded with NAs as needed.
}
\description{
This functions is similar to cbind in that it binds xts matrices column-wise.
If the xts matrices do not have the same number of rows, it will appropriately
pad with NAs.  In addition, it binds on date (not time), so if the prices
are sampled at different times, then the times are ignored for proper binding.
}

