% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{xtscagr}
\alias{xtscagr}
\title{Compute the annualized return from an xts of equity curves (prices)}
\usage{
xtscagr(ec)
}
\arguments{
\item{ec}{An xts matrix of equity curves expressed as prices.}
}
\value{
Returns a data frame of CAGR
}
\description{
The equity curves are in the form of prices, not returns.
}

