% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{xtsmdd}
\alias{xtsmdd}
\title{Compute the maximum drawdown of an xts matrix of returns.}
\usage{
xtsmdd(rets, digits = 4)
}
\arguments{
\item{rets}{An xts matrix of equity curves, expressed as periodic returns}

\item{digits}{Number of digits to report}
}
\value{
Returns a dataframe of maximum drawdowns, expressed as fractions.
}
\description{
Returns are assumed to be geometric ("discrete").  This function
is very fast and works on wide xts matrices (many equity curves)
}

