% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{xtsnormalize}
\alias{xtsnormalize}
\title{Normalizes an xts matrix agaist its first row.}
\usage{
xtsnormalize(xtsmat, na.rm = TRUE)
}
\arguments{
\item{xtsmat}{The xts matrix to normalize.}

\item{na.rm}{Logical.  Should we remove all rows with NAs?}
}
\value{
Returns an xts matrix that is normalized against its own
        first row.
}
\description{
Normalizes an xts matrix agaist its first row.
}
