% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsplot.R
\name{make_colors}
\alias{make_colors}
\title{Make a nice color palette appropriate for plotting functions}
\usage{
make_colors(n = 0, type = c("lines", "regions"), showcolors = FALSE)
}
\arguments{
\item{n}{Number of colors to return.  If n is larger than the palette
selected, then the colors are recycled.}

\item{type}{The type of color palette to create.  If type = 'lines'
(the default), then the palette is a set of colors most
appropriate to plot lines on a white background.
If type = 'regions' then the palette is a set of
pastel colors most appropriate to highlight regions on
a plot or to make a bar graph.}

\item{showcolors}{Logical.  Specifies whether to plot a bar chart that shows
the colors in the selected palette with their names.}
}
\value{
Returns a named vector of RGB colors (in hex format).
        xtsanalytics functions can refer to the names of these colors rather
        than their index number when the user specifies a given set of colors.
        See xtsplot for details.
}
\description{
Make a nice color palette appropriate for plotting functions
}
