% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxscreen.R
\name{maxscreen}
\alias{maxscreen}
\title{Volatility adjusted momentum screen for optimizing portfolios}
\usage{
maxscreen(prices, maxweights = 0.5, volfeature = "sd252",
  volthresh = 0.12, momfeature = "mom252", momthresh = 0.05,
  cashasset = 1, rankthresh = ncol(prices))
}
\arguments{
\item{prices}{An xts matrix of asset prices.}

\item{maxweights}{A vector containing the maximum allowable weight
for each asset. The vector is recycled or truncated
if too short or too long.  Default is 0.5.}

\item{volfeature}{The volatility feature that will be computed from
the prices matrix using function make_feature.
Default is sd252.}

\item{volthresh}{The volatility threshold at which the asset weight
will be scaled back, according to the formula
min(1, volthresh / asset_volatility), where
asset_volatility is the result of the volfeature
feature.}

\item{cashasset}{The column number or name of the cash asset.  This
asset has a fixed weight of 1 so the optimizer can
always select it, if all other assets have a zero
weight.  Default is column 1.}

\item{rankthres}{The rank value beyond which an asset weight is forced
to zero.  For example, if rankthres = 4, then the top
4 highest momentum asset are included at the given date
while all other assets are excluded by setting their
weights to zero.  The cash asset however is always included
and treated separately.  So in the above example, the top
4 assets are included plus the cash asset.}
}
\value{
Returns an xts matrix of asset weights. Some weights may be
        zero if the asset was screened out at the given time index.
}
\description{
Returns an xts matrix of asset weights, modulated according to
a set of rules.
}
\details{
The xts matrix returned will generally start at a later time
that the prices matrix since features are calculated from it.
Therefore, you must ensure that it properly aligns with
other matrices in the calling function.

The first rule applied is...
}

