% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskparity_regularize.R
\name{riskparity_regularize}
\alias{riskparity_regularize}
\title{Screen assets and adjust max portfolio weights using risk parity regularization}
\usage{
riskparity_regularize(prices, maxweights = 0.5, volfeature = "sd126",
  riskparity = 0.12, momfeature = "mom252", momthresh = 0.05,
  cashasset = 1, Nassets = ncol(prices))
}
\arguments{
\item{prices}{An xts matrix of asset prices.}

\item{maxweights}{A vector containing the maximum allowable weightfor each
asset as a fraction of the total portfolio. The vector
is recycled if not enough values are provided, or truncated
if it is too long.  Default is 0.5 for all assets.}

\item{volfeature}{The volatility feature used to calculate the risk parity
maximum weight adjustment.  Risk parity in this context is
used to equalize risk for a given asset when compared
against its stated maximum weight.  For example, if an
asset has its maxweight set to 0.3, its risk parity
threshold set to 0.10 and its volatility at a given time
is 0.15.  Then its adjusted maxweight will be
0.3 * (0.10 / 0.15) = 0.20. Default is sd126.}

\item{riskparity}{The volatility threshold at which the asset weight
will be scaled back, according to the formula
min(1, riskparity / asset_volatility), where
asset_volatility is the current result of the volfeature
feature, expressed as an annualized value.}

\item{momfeature}{}

\item{momthresh}{}

\item{cashasset}{The column number or name of the cash asset.  This
asset has a fixed weight of 1 so the optimizer can
always select it, if all other assets have a zero
weight.  Default is column 1.}

\item{Nassets}{The maximum number of assets in the portfolio at any
                   given time, including the cash asset.  This is used
                   to derive a rank threshold value beyond to screen the
                   top assets at each WFO date.


                   to zero.  For example, if rankthres = 4, then the top
                   4 highest momentum asset are included at the given date
                   while all other assets are excluded by setting their
                   weights to zero.  The cash asset however is always included
                   and treated separately.  So in the above example, the top
                   4 assets are included plus the cash asset.}
}
\value{
Returns a list of two xts matrices:  the first xts "maxweights" contains
                the maximum asset weights, where the screened out asset weights
                are set to zero. The second xts contains the adjusted momentum scores
                (annualized) according to the following regularization function:
}
\description{
This function is used to screen assets from a universe and adjust
the weights of each assets according to certain risk parity rules,
momentum and portfolio correlation.
}
\details{
Returns an xts matrix of asset weights, modulated according to
a set of rules.

The xts matrix returned will generally start at a later time
that the prices matrix since features are calculated from it.
Therefore, you must ensure that it properly aligns with
other matrices in the calling function.

The first rule applied is...
}

