% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpurpose.R
\name{sprint}
\alias{sprint}
\title{Print to console using C-like sprintf formatting.}
\usage{
sprint(string, ...)
}
\arguments{
\item{string}{A character string containing text and formatting style.}

\item{...}{Additional parameters passed to sprintf.  In most cases, these
are the sequence of variables that are printed, provided they
have an associated formatting in the string.}
}
\value{
No return value.  This function is called for its side effects
        by printing on the console.
}
\description{
A simple function to print text and variable content on the console, using
the C-like function sprintf formatting style.  Also automatically appends
a carriage return at the end of the string.
}
\seealso{
sprintf()
}

